feat: reorganize folders structure
