+++ web/dashboard/src/app/(session)/components/generate/page.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { NewProjectModalPage } from './Chat';
-import { AI } from "@/data/ai/actions";
+import { NewComponent } from '../../../../components/components/create/CreateNewComponent';
+import { AI } from "@/components/components/create/utils/actions";
 
 export default async function GenerateComponentPage() {
     return (
         <AI>
             <div className="container h-full flex-1 flex-col space-y-8 md:flex" style={{ height: "calc(100vh - 10rem)"}}>
-                <NewProjectModalPage />
+                <NewComponent />
             </div>
         </AI>
     );
