+++ web/dashboard/src/components/components/create/generatedCodeDisplay.tsx
