+++ web/dashboard/src/components/components/create/utils/chat-schema.ts
