+++ web/dashboard/src/components/components/create/utils/use-form.ts
@@ -1,5 +1,5 @@
 import { zodResolver } from "@hookform/resolvers/zod";
-import { type FieldValues, useForm as useHookForm } from "react-hook-form";
+import { useForm as useHookForm } from "react-hook-form";
 import { chatSchema } from "./chat-schema";
 
 export const useForm = <T>() => {
