+++ web/site/src/content/docs/reference/components.mdx
@@ -37,7 +37,7 @@ export const sampleDialog = `{
 
 ### Component structure
 
-Example of a Hello World Component - [View the Code](https://github.com/diconium/ethereal-nexus/tree/6b0145b1cb33d84047ce7e0f23569a106f6dc64c/example-vite-react-ts/src/components/ethereal-nexus/ReactHelloWorl)
+Example of a Hello World Component - [View the Code](https://github.com/diconium/ethereal-nexus/tree/6b0145b1cb33d84047ce7e0f23569a106f6dc64c/example-vite-react-ts/src/components/ReactHelloWorld)
 
 * HelloWorld.tsx
 * HelloWorld.config.json
